discord-chatbot/requirements.txt
python-dotenv==1.0.1
langchain_openai==0.3.17
langchain==0.3.25
langchain_community==0.3.24
langgraph==0.4.5
pymongo==4.12.1
discord.py==2.5.2
streamlit==1.45.1
langfuse==2.60.5
httpx==0.28.1
tiktoken==0.9.0
